chore(iam): moved common code to utils files
