+++ iam/src/index.ts
@@ -18,16 +18,10 @@ import {
   ChallengeRequestBody,
   VerifyRequestBody,
   CredentialResponseBody,
-  ProviderContext,
   CheckRequestBody,
   EasPayload,
   PassportAttestation,
   EasRequestBody,
-  VerifiedPayload,
-  PROVIDER_ID,
-  VerifiableCredential,
-  ValidResponseBody,
-  SignatureType,
 } from "@gitcoin/passport-types";
 import onchainInfo from "../../deployments/onchainInfo.json" assert { type: "json" };
 
@@ -36,22 +30,21 @@ import { getEASFeeAmount } from "./utils/easFees.js";
 import * as stampSchema from "./utils/easStampSchema.js";
 import * as passportSchema from "./utils/easPassportSchema.js";
 import { hasValidIssuer, getIssuerKey } from "./issuers.js";
+import { checkConditionsAndIssueCredentials, verifyTypes } from "./utils/credentials.js";
 
 // ---- Generate & Verify methods
 import * as DIDKit from "@spruceid/didkit-wasm-node";
-import { issueChallengeCredential, issueHashedCredential, verifyCredential } from "@gitcoin/passport-identity";
+import { issueChallengeCredential, verifyCredential } from "@gitcoin/passport-identity";
 
 // All provider exports from platforms
-import { providers, platforms } from "@gitcoin/passport-platforms";
 
 import path from "path";
 import { fileURLToPath } from "url";
-import { IAMError } from "./utils/scorerService.js";
 import { VerifyDidChallengeBaseError } from "./utils/verifyDidChallenge.js";
-import { errorRes } from "./utils/helpers.js";
+import { errorRes, addErrorDetailsToMessage, ApiError } from "./utils/helpers.js";
 import { ATTESTER_TYPES, getAttestationDomainSeparator, getAttestationSignerForChain } from "./utils/attestations.js";
 import { scrollDevBadgeHandler } from "./utils/scrollDevBadge.js";
-import axios from "axios";
+import { autoVerificationHandler } from "./utils/autoVerification.js";
 
 // ---- Config - check for all required env variables
 // We want to prevent the app from starting with default values or if it is misconfigured
@@ -112,33 +105,6 @@ if (configErrors.length > 0) {
 
 const EAS_FEE_USD = parseFloat(process.env.EAS_FEE_USD);
 
-const providerTypePlatformMap = Object.entries(platforms).reduce(
-  (acc, [platformName, { providers }]) => {
-    providers.forEach(({ type }) => {
-      acc[type] = platformName;
-    });
-
-    return acc;
-  },
-  {} as { [k: string]: string }
-);
-
-function groupProviderTypesByPlatform(types: string[]): string[][] {
-  return Object.values(
-    types.reduce(
-      (groupedProviders, type) => {
-        const platform = providerTypePlatformMap[type] || "generic";
-
-        if (!groupedProviders[platform]) groupedProviders[platform] = [];
-        groupedProviders[platform].push(type);
-
-        return groupedProviders;
-      },
-      {} as { [k: keyof typeof platforms]: string[] }
-    )
-  );
-}
-
 // create the app and run on port
 export const app = express();
 
@@ -147,75 +113,6 @@ app.use(express.json());
 
 // set cors to accept calls from anywhere
 app.use(cors());
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const addErrorDetailsToMessage = (message: string, error: any): string => {
-  if (error instanceof IAMError || error instanceof Error) {
-    message += `, ${error.name}: ${error.message}`;
-  } else if (typeof error === "string") {
-    message += `, ${error}`;
-  }
-  return message;
-};
-
-// return response for given payload
-const issueCredentials = async (
-  types: string[],
-  address: string,
-  payload: RequestPayload
-): Promise<CredentialResponseBody[]> => {
-  // if the payload includes an additional signer, use that to issue credential.
-  if (payload.signer) {
-    // We can assume that the signer is a valid address because the challenge was verified within the /verify endpoint
-    payload.address = payload.signer.address;
-  }
-
-  const results = await verifyTypes(types, payload);
-
-  return await Promise.all(
-    results.map(async ({ verifyResult, code: verifyCode, error: verifyError, type }) => {
-      let code = verifyCode;
-      let error = verifyError;
-      let record, credential;
-
-      try {
-        // check if the request is valid against the selected Identity Provider
-        if (verifyResult.valid === true) {
-          // construct a set of Proofs to issue a credential against (this record will be used to generate a sha256 hash of any associated PII)
-          record = {
-            // type and address will always be known and can be obtained from the resultant credential
-            type: verifyResult.record.pii ? `${type}#${verifyResult.record.pii}` : type,
-            // version is defined by entry point
-            version: "0.0.0",
-            // extend/overwrite with record returned from the provider
-            ...(verifyResult?.record || {}),
-          };
-
-          const currentKey = getIssuerKey(payload.signatureType);
-          // generate a VC for the given payload
-          ({ credential } = await issueHashedCredential(
-            DIDKit,
-            currentKey,
-            address,
-            record,
-            verifyResult.expiresInSeconds,
-            payload.signatureType
-          ));
-        }
-      } catch {
-        error = "Unable to produce a verifiable credential";
-        code = 500;
-      }
-
-      return {
-        record,
-        credential,
-        code,
-        error,
-      };
-    })
-  );
-};
 
 // health check endpoint
 app.get("/health", (_req, res) => {
@@ -307,143 +204,6 @@ app.post("/api/v0.0.0/check", (req: Request, res: Response): void => {
     .catch(() => errorRes(res, "Unable to check payload", 500));
 });
 
-type VerifyTypeResult = {
-  verifyResult: VerifiedPayload;
-  type: string;
-  error?: string;
-  code?: number;
-};
-
-export async function verifyTypes(types: string[], payload: RequestPayload): Promise<VerifyTypeResult[]> {
-  // define a context to be shared between providers in the verify request
-  // this is intended as a temporary storage for providers to share data
-  const context: ProviderContext = {};
-  const results: VerifyTypeResult[] = [];
-
-  await Promise.all(
-    // Run all platforms in parallel
-    groupProviderTypesByPlatform(types).map(async (platformTypes) => {
-      // Iterate over the types within a platform in series
-      // This enables providers within a platform to reliably share context
-      for (let type of platformTypes) {
-        let verifyResult: VerifiedPayload = { valid: false };
-        let code, error;
-
-        const realType = type;
-        if (type.startsWith("AllowList")) {
-          payload.proofs = {
-            ...payload.proofs,
-            allowList: type.split("#")[1],
-          };
-          type = "AllowList";
-        } else if (type.startsWith("DeveloperList")) {
-          // Here we handle the custom DeveloperList stamps
-          const [_type, conditionName, conditionHash, ..._rest] = type.split("#");
-          payload.proofs = {
-            ...payload.proofs,
-            conditionName,
-            conditionHash,
-          };
-          type = "DeveloperList";
-        }
-
-        try {
-          // verify the payload against the selected Identity Provider
-          verifyResult = await providers.verify(type, payload, context);
-          if (!verifyResult.valid) {
-            code = 403;
-            // TODO to be changed to just verifyResult.errors when all providers are updated
-            const resultErrors = verifyResult.errors;
-            error = resultErrors?.join(", ")?.substring(0, 1000) || "Unable to verify provider";
-            if (error.includes(`Request timeout while verifying ${type}.`)) {
-              console.log(`Request timeout while verifying ${type}`);
-              // If a request times out exit loop and return results so additional requests are not made
-              break;
-            }
-          }
-          if (type === "AllowList") {
-            type = `AllowList#${verifyResult.record.allowList}`;
-          } else {
-            type = realType;
-          }
-        } catch (e) {
-          error = "Unable to verify provider";
-          code = 400;
-        }
-
-        results.push({ verifyResult, type, code, error });
-      }
-    })
-  );
-
-  return results;
-}
-
-const verifyAdditionalSigner = async ({
-  challenge,
-  signature,
-  address,
-}: {
-  challenge: VerifiableCredential;
-  signature: string;
-  address: string;
-}): Promise<{ verifiedAddress: string }> => {
-  const additionalSignerCredential = await verifyCredential(DIDKit, challenge);
-  const verifiedAddress = utils.getAddress(utils.verifyMessage(challenge.credentialSubject.challenge, signature));
-
-  if (!additionalSignerCredential || verifiedAddress.toLowerCase() !== address.toLowerCase()) {
-    throw new Error("Unable to verify payload signer");
-  }
-
-  return { verifiedAddress };
-};
-
-const checkConditionsAndIssueCredentials = async (
-  payload: RequestPayload,
-  address: string
-): Promise<{
-  credentials?: CredentialResponseBody[];
-  error?: {
-    message: string;
-    code: number;
-  };
-}> => {
-  // Verify additional signer if provided
-  if (payload.signer) {
-    const { verifiedAddress } = await verifyAdditionalSigner(payload.signer);
-    payload.signer.address = verifiedAddress;
-  }
-
-  const singleType = !payload.types?.length;
-  const types = (!singleType ? payload.types : [payload.type]).filter((type) => type);
-
-  // Validate requirements and issue credentials
-  if (payload && payload.type) {
-    const responses = await issueCredentials(types, address, payload);
-
-    if (singleType) {
-      const response = responses[0];
-      if ("error" in response && response.code && response.error) {
-        return {
-          error: {
-            message: response.error,
-            code: response.code,
-          },
-        };
-      }
-      return { credentials: [response] };
-    }
-    return { credentials: responses };
-  }
-
-  return {
-    error: {
-      message: "Invalid payload",
-      code: 400,
-    },
-  };
-};
-
 // expose verify entry point
 // verify a users claim to stamps and issue the stamps if the claim is valid
 app.post("/api/v0.0.0/verify", (req: Request, res: Response): void => {
@@ -483,17 +243,16 @@ app.post("/api/v0.0.0/verify", (req: Request, res: Response): void => {
 
         const result = await checkConditionsAndIssueCredentials(payload, address);
 
-        if (result.error) {
-          return void errorRes(res, result.error.message, result.error.code);
-        }
-
-        return void res.json(result.credentials);
+        return void res.json(result);
       }
 
       // error response
       return void errorRes(res, "Unable to verify payload", 401);
     })
     .catch((error) => {
+      if (error instanceof ApiError) {
+        return void errorRes(res, error.message, error.code);
+      }
       let message = "Unable to verify payload";
       if (error instanceof Error) message += `: ${error.name}`;
       return void errorRes(res, message, 500);
@@ -718,86 +477,7 @@ app.post("/api/v0.0.0/eas/score", async (req: Request, res: Response) => {
   }
 });
 
-type AutoVerificationRequest = {
-  address: string;
-  scorerId: string;
-};
-
-const apiKey = process.env.SCORER_API_KEY;
-
-// expose verify entry point
-// verify a users claim to stamps and issue the stamps if the claim is valid
-app.post("/api/v0.0.0/auto-verification", async (req: Request, res: Response): Promise<void> => {
-  try {
-    const { address, scorerId } = req.body as AutoVerificationRequest;
-
-    const evmPlatforms = Object.values(platforms).filter(({ PlatformDetails }) => PlatformDetails.isEVM);
-
-    if (!utils.isAddress(address)) {
-      return void errorRes(res, "Invalid address", 400);
-    }
-
-    // TODO we should use the scorerId to check for any EVM stamps particular to a community, and include those
-    // here. We'll have to add an endpoint for this
-
-    const evmProviders: PROVIDER_ID[] = evmPlatforms
-      .map(({ ProviderConfig }) => ProviderConfig.map(({ providers }) => providers.map(({ name }) => name)))
-      .flat(2);
-
-    console.log("evmProviders", evmProviders);
-
-    const payload = {
-      address,
-      type: "EVMBulkVerify",
-      types: evmProviders,
-      version: "0.0.0",
-      signatureType: "EIP712" as SignatureType,
-    };
-
-    const result = await checkConditionsAndIssueCredentials(payload, address);
-
-    if (result.error) {
-      return void errorRes(res, result.error.message, result.error.code);
-    }
-
-    const validCredentials = (result.credentials || [])
-      .filter((cred): cred is ValidResponseBody => (cred as ValidResponseBody).credential !== undefined)
-      .map(({ credential }) => credential);
-
-    const scorerResponse: {
-      data?: {
-        score?: {
-          score?: string;
-          evidence?: {
-            rawScore?: string;
-          };
-        };
-      };
-    } = await axios.post(
-      `${process.env.SCORER_ENDPOINT}/internal/stamps/${address}`,
-      {
-        stamps: validCredentials,
-        scorer_id: scorerId,
-      },
-      {
-        headers: {
-          Authorization: apiKey,
-        },
-      }
-    );
-
-    const scoreData = scorerResponse.data?.score || {};
-
-    const score = String(scoreData.evidence?.rawScore || scoreData.score);
-
-    // TODO should we issue a score VC?
-
-    return void res.json({ score });
-  } catch (error) {
-    const message = addErrorDetailsToMessage("Unexpected error when processing request", error);
-    return void errorRes(res, message, 500);
-  }
-});
+app.post("/api/v0.0.0/auto-verification", autoVerificationHandler);
 
 // procedure endpoints
 app.use("/procedure", procedureRouter);
