+++ iam/src/utils/helpers.ts
@@ -1,5 +1,23 @@
 import { Response } from "express";
+import { IAMError } from "./scorerService.js";
 
 // return a JSON error response with a 400 status
 export const errorRes = (res: Response, error: string | object, errorCode: number): Response =>
   res.status(errorCode).json({ error });
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const addErrorDetailsToMessage = (message: string, error: any): string => {
+  if (error instanceof IAMError || error instanceof Error) {
+    message += `, ${error.name}: ${error.message}`;
+  } else if (typeof error === "string") {
+    message += `, ${error}`;
+  }
+  return message;
+};
+
+export class ApiError extends Error {
+  constructor(public message: string, public code: number) {
+    super(message);
+    this.name = this.constructor.name;
+  }
+}
